refactor(plugins): give each default plugin a package.json manifest
Move each first-party default from a flat defaults/<name>.ts into
defaults/<name>/register.ts paired with a package.json (name + version +
peerDependencies["@vellumai/plugin-api"]), and derive the in-code manifest
from that package.json (name: pkg.name, version: pkg.version). bun --compile
bundles the JSON imports, so the defaults stay baked into the binary.

This aligns the defaults with the external-plugin layout the docs describe
and backs plugins.md TL;DR step 2; the register.ts entry is the interim shim
to be stripped toward file-based discovery in later PRs.
