+++ assistant/src/plugins/defaults/compaction/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "default-compaction",
+  "version": "1.0.0",
+  "description": "First-party default plugin wrapping the assistant's built-in compaction pipeline with a passthrough implementation.",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "main": "./register.ts",
+  "engines": {
+    "node": ">=20.12.0"
+  },
+  "peerDependencies": {
+    "@vellumai/plugin-api": "^0.8.0"
+  }
+}
