+++ assistant/src/plugins/defaults/memory-retrieval/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "default-memory-retrieval",
+  "version": "0.0.1",
+  "description": "First-party default plugin wrapping the assistant's built-in memory-retrieval pipeline with a passthrough implementation.",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "main": "./register.ts",
+  "engines": {
+    "node": ">=20.12.0"
+  },
+  "peerDependencies": {
+    "@vellumai/plugin-api": "^0.8.0"
+  }
+}
