+++ assistant/src/plugins/defaults/tool-result-truncate/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "default-tool-result-truncate",
+  "version": "1.0.0",
+  "description": "First-party default plugin wrapping the assistant's built-in tool-result-truncate pipeline with a passthrough implementation.",
+  "private": true,
+  "license": "MIT",
+  "type": "module",
+  "main": "./register.ts",
+  "engines": {
+    "node": ">=20.12.0"
+  },
+  "peerDependencies": {
+    "@vellumai/plugin-api": "^0.8.0"
+  }
+}
